[refactor] - splits up test_query to test_query (internal) and test_gamemanager for travis testing
